requirements.txt
# Trimmed to the dependencies directly used by the current manuscript code paths.
# Tested against Python 3.12; keep the torch stack aligned for your platform.

h5py>=3.15.1
imagecodecs>=2026.1.14
lighter_zoo
matplotlib>=3.10.8
nibabel>=5.4.2
numpy>=2.4.4
openslide-python  # only needed for src/extractors/wsi_slide_ops.py; requires native OpenSlide
pandas>=3.0.0
Pillow>=12.1.1
PyYAML>=6.0.3
scipy>=1.17.0
seaborn>=0.13.2  # optional in train.py, but improves plot palettes
tensorboard>=2.20.0
timm>=1.0.25
torch>=2.10.0
torch-geometric>=2.7.0
torchvision>=0.25.0
tqdm>=4.67.3
